requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-cosmos
requests
azure-ai-translation-text
azure-ai-translation-document==1.0.0
azure-ai-textanalytics
openai==0.27.0
